Support for >2 nested subgroups in destination path
Pulling an object out of a gitlab.v4.objects.Group's subgroup property returns a gitlab.v4.objects.GroupSubgroup object, which is not a subclass of Group, and thus doesn't have a subgroup property.  This will cause the code to break if the destination is more than 2 groups deep.

To solve this the id property of the Subgroup can be used to pull out an actual group object object out with groups.get, then querying the subgroups of that object.